Move tests to one file
